rider mark availabel and Inactivate add to redis

+++ src/main/java/com/example/smartbite/store/rider_modul/DTO/RiderAvailableDTO.java
@@ -4,7 +4,9 @@ import java.util.UUID;
 
 public record RiderAvailableDTO(
         UUID rider_id ,
-        Float longitude,
-        Float latitude
+        Double longitude,
+        Double latitude
+//        Float longitude,
+//        Float latitude
 ) {
 }
